LPS-120270 Move logic to AuthVerifierFilter
